# Use RecyclableMemoryStream instead of MemoryStream