What is done on this PR:

* Make use of a `RecyclableMemoryStreamManager` from the `Microsoft.IO.RecyclableMemoryStream` to recycle `MemoryStream` instead of having extensive use of the Garbage Collector. More info here: https://github.com/microsoft/Microsoft.IO.RecyclableMemoryStream

I have a benchmark in place for a project I am working on. Here are the result benchmarks.

Before:

|                Method |       Mean |    Error |   StdDev |    Gen0 |    Gen1 |  Allocated |
|---------------------- |-----------:|---------:|---------:|--------:|--------:|-----------:|
| HttpWithClientFactory | 1,036.2 μs |  8.33 μs |  7.79 μs | 15.6250 |  3.9063 |  253.87 KB |
|                WsText | 4,626.6 μs | 49.26 μs | 43.67 μs | 39.0625 | 15.6250 | 2675.36 KB |
|              WsBinary |         NA |       NA |       NA |      NA |      NA |         NA |

After:

|                Method |       Mean |    Error |   StdDev |    Gen0 |   Gen1 |  Allocated |
|---------------------- |-----------:|---------:|---------:|--------:|-------:|-----------:|
| HttpWithClientFactory | 1,033.3 us |  4.92 us |  4.37 us | 15.6250 | 3.9063 |  253.67 KB |
|                WsText | 3,212.6 us | 60.18 us | 53.34 us | 27.3438 | 7.8125 | 1166.44 KB |
|              WsBinary |         NA |       NA |       NA |      NA |     NA |         NA |

I put the `HttpWithClientFactory` in reference because it should have a very close performance impact to `WsText`. CPU time and memory allocation are reduced by 80%-120%. Of course, it depends on the use case. For some scenarii, there are absolutely no impact (small batch, few/no data consumption).