@Odonno 2023-08-08
Unfortunately, those 2 concepts do not work easily together on this context. Either 
* we consume a `Memory<byte>` from a `RecycledMemoryStream` and `Dispose` it (when disposing the ResponseMessage), so it changes the API a little
* or we use a `RecycledMemoryStream` as it is and export a `byte[]` (without the `Memory<byte>` then), potentially losing the gain on memory allocation...