@Odonno 2023-08-12
Some notes:

* ResponseMessage now has a MemoryStream property
* Dispose of the Stream need to be handled manually for this so we can access the stream
* A new property on the client: IsStreamDisposedAutomatically, allows to handle the dispose manually to have access to inner MemoryStream...
* ...default value is true, stream will be disposed automatically/immediately and so we do not provide access to the MemoryStream (as it is disposed before access)
* refactor the code around ReceiveClient to fully use the RecyclableMemoryStream & use recommended Memory<byte> instead of ArraySegment<byte> (except for .NET Standard 2.0) => I did not perceive any significant impact in terms of performance/memory allocation doing that, and the code is way cleaner 

I made some final touches and now it is good for review.