@Odonno 2023-08-12
Just a final final touch:

* Add `SendTextAsBinary` and `SendTextAsBinaryAsync` methods to send TEXT message without having to serialize again from `string` to `byte[]`. On benchmark, it reduces memory allocation by half.