@Odonno 2023-09-18
Hi @Marfusios 

Thank you for your effort. I resolved the conflicts after your changes. Mostly nullable reference types. Not sure if I am 100% compliant on this topic. Will make some check later to see it works as expected.