@Odonno 2023-09-24
Since you upgraded to .NET Standard 2.1, I removed pragma conditions for .NET Standard 2.0.