@Marfusios 2024-02-14
Hey, finally got to it, thanks for the PR, looks good to me from here, will test it locally after the merge and do some slight polishing